Extracted items from `model.ts` to `internalTreeModel.ts`. - Renamed `model.ts` to `pageModel.ts` - Extracted common items from `tree.ts` to `../tree.ts`.
