+++ src/page/internalTree/tree.ts
@@ -2,16 +2,14 @@
  * Contains common items.
  */
 
+import { Color, PageContent, PageContentMutable } from "../pageModel";
 import {
   BufferCursor,
   CharValues,
-  Color,
+  InternalTreeNode,
+  InternalTreeNodeMutable,
   NEWLINE,
-  Node,
-  NodeMutable,
-  PageContent,
-  PageContentMutable,
-} from "../model";
+} from "./internalTreeModel";
 
 /**
  * The maximum length of a buffer string.
@@ -25,7 +23,7 @@ export interface NodePositionOffset {
   /**
    * Piece Index
    */
-  readonly node: Node;
+  readonly node: InternalTreeNode;
 
   /**
    * The index of the node inside the array.
@@ -44,14 +42,6 @@ export interface NodePositionOffset {
 }
 
 /**
- * Contains a node and its index in a page/piece table.
- */
-export interface NodePosition {
-  readonly node: Node;
-  readonly index: number;
-}
-
-/**
  * Finds the node which contains the offset.
  * @param offset The offset.
  * @param nodes The nodes.
@@ -59,7 +49,7 @@ export interface NodePosition {
  */
 export function findNodeAtOffset(
   offset: number,
-  nodes: ReadonlyArray<Node>,
+  nodes: ReadonlyArray<InternalTreeNode>,
   root: number,
 ): NodePositionOffset {
   let x = root;
@@ -229,8 +219,8 @@ export function recomputeTreeMetadata(
   lineFeedDelta =
     calculateLineFeedCount(page, page.nodes[x].left) -
     page.nodes[x].leftLineFeedCount;
-  (page.nodes[x] as NodeMutable).leftCharCount += lengthDelta;
-  (page.nodes[x] as NodeMutable).leftLineFeedCount += lineFeedDelta;
+  (page.nodes[x] as InternalTreeNodeMutable).leftCharCount += lengthDelta;
+  (page.nodes[x] as InternalTreeNodeMutable).leftLineFeedCount += lineFeedDelta;
 
   // go upwards till root. O(logN)
   while (x !== page.root && (lengthDelta !== 0 || lineFeedDelta !== 0)) {
@@ -240,10 +230,10 @@ export function recomputeTreeMetadata(
       };
       (page.nodes[
         page.nodes[x].parent
-      ] as NodeMutable).leftCharCount += lengthDelta;
+      ] as InternalTreeNodeMutable).leftCharCount += lengthDelta;
       (page.nodes[
         page.nodes[x].parent
-      ] as NodeMutable).leftLineFeedCount += lineFeedDelta;
+      ] as InternalTreeNodeMutable).leftLineFeedCount += lineFeedDelta;
     }
 
     x = page.nodes[x].parent;
@@ -296,7 +286,7 @@ export const SENTINEL_INDEX = 0;
 /**
  * The sentinel node of red-black trees.
  */
-export const SENTINEL: Node = {
+export const SENTINEL: InternalTreeNode = {
   bufferIndex: 0,
   start: { column: 0, line: 0 },
   end: { column: 0, line: 0 },
@@ -316,42 +306,18 @@ export const SENTINEL: Node = {
  * @param page The page/piece table which contains the `SENTINEL` node.
  */
 export function resetSentinel(page: PageContentMutable): void {
-  (SENTINEL as NodeMutable).bufferIndex = 0;
-  (SENTINEL as NodeMutable).start = { column: 0, line: 0 };
-  (SENTINEL as NodeMutable).end = { column: 0, line: 0 };
-  (SENTINEL as NodeMutable).leftCharCount = 0;
-  (SENTINEL as NodeMutable).leftLineFeedCount = 0;
-  (SENTINEL as NodeMutable).length = 0;
-  (SENTINEL as NodeMutable).lineFeedCount = 0;
-  (SENTINEL as NodeMutable).color = Color.Black;
-  (SENTINEL as NodeMutable).parent = SENTINEL_INDEX;
-  (SENTINEL as NodeMutable).left = SENTINEL_INDEX;
-  (SENTINEL as NodeMutable).right = SENTINEL_INDEX;
+  (SENTINEL as InternalTreeNodeMutable).bufferIndex = 0;
+  (SENTINEL as InternalTreeNodeMutable).start = { column: 0, line: 0 };
+  (SENTINEL as InternalTreeNodeMutable).end = { column: 0, line: 0 };
+  (SENTINEL as InternalTreeNodeMutable).leftCharCount = 0;
+  (SENTINEL as InternalTreeNodeMutable).leftLineFeedCount = 0;
+  (SENTINEL as InternalTreeNodeMutable).length = 0;
+  (SENTINEL as InternalTreeNodeMutable).lineFeedCount = 0;
+  (SENTINEL as InternalTreeNodeMutable).color = Color.Black;
+  (SENTINEL as InternalTreeNodeMutable).parent = SENTINEL_INDEX;
+  (SENTINEL as InternalTreeNodeMutable).left = SENTINEL_INDEX;
+  (SENTINEL as InternalTreeNodeMutable).right = SENTINEL_INDEX;
   page.nodes[0] = SENTINEL;
-}
-
-/**
- * Finds the minimum of the subtree given by the `x`
- * @param page The piece table/page.
- * @param x The index from which to find the minimum of that subtree.
- */
-export function treeMinimum(page: PageContent, x: number): NodePosition {
-  while (page.nodes[x].left !== SENTINEL_INDEX) {
-    x = page.nodes[x].left;
-  }
-  return { node: page.nodes[x], index: x };
-}
-
-/**
- * Finds the maximum of the subtree given by the `x`
- * @param page The piece table/page.
- * @param x The index from which to find the maximum of that subtree.
- */
-export function treeMaximum(page: PageContent, x: number): NodePosition {
-  while (page.nodes[x].right !== SENTINEL_INDEX) {
-    x = page.nodes[x].right;
-  }
-  return { node: page.nodes[x], index: x };
 }
 
 /**
@@ -381,61 +347,5 @@ export function updateTreeMetadata(
     }
 
     x = page.nodes[x].parent;
-  }
-}
-
-/**
- * Gets the next node of a red-black tree, given the current node's index.
- * @param page The page/piece table.
- * @param currentNode The index of the current node in the `page.nodes` array.
- */
-export function nextNode(page: PageContent, currentNode: number): NodePosition {
-  if (page.nodes[currentNode].right !== SENTINEL_INDEX) {
-    return treeMinimum(page, page.nodes[currentNode].right);
-  }
-
-  while (page.nodes[currentNode].parent !== SENTINEL_INDEX) {
-    if (page.nodes[page.nodes[currentNode].parent].left === currentNode) {
-      break;
-    }
-
-    currentNode = page.nodes[currentNode].parent;
-  }
-
-  if (page.nodes[currentNode].parent === SENTINEL_INDEX) {
-    return { node: SENTINEL, index: SENTINEL_INDEX };
-  } else {
-    return {
-      index: page.nodes[currentNode].parent,
-      node: page.nodes[page.nodes[currentNode].parent],
-    };
-  }
-}
-
-/**
- * Gets the previous node of a red-black tree, given the current node's index.
- * @param page The page/piece table.
- * @param currentNode The index of the current node in the `page.nodes` array.
- */
-export function prevNode(page: PageContent, currentNode: number): NodePosition {
-  if (page.nodes[currentNode].left !== SENTINEL_INDEX) {
-    return treeMaximum(page, page.nodes[currentNode].left);
-  }
-
-  while (page.nodes[currentNode].parent !== SENTINEL_INDEX) {
-    if (page.nodes[page.nodes[currentNode].parent].right === currentNode) {
-      break;
-    }
-
-    currentNode = page.nodes[currentNode].parent;
-  }
-
-  if (page.nodes[currentNode].parent === SENTINEL_INDEX) {
-    return { node: SENTINEL, index: SENTINEL_INDEX };
-  } else {
-    return {
-      index: page.nodes[currentNode].parent,
-      node: page.nodes[page.nodes[currentNode].parent],
-    };
   }
 }
